backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
pdf2image==1.17.0
pdfplumber==0.11.0
pdf2docx==0.5.8
python-docx==1.1.0
python-pptx==0.6.23
pytesseract==0.3.10
pdfkit==1.0.0
python-magic==0.4.27
aiofiles==23.2.1
slowapi==0.1.9
